+++ src/main/java/com/synclab/ecommerce/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package com.synclab.ecommerce.repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.synclab.ecommerce.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 	Optional<User> findByFirstName(String name);
