+++ src/main/java/com/synclab/ecommerce/service/UserService.java
@@ -1,6 +1,5 @@
 package com.synclab.ecommerce.service;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 import com.synclab.ecommerce.model.User;
@@ -9,7 +8,7 @@ public interface UserService {
 
 	Optional<User> findByFirstName(String name);
 
-	Optional<User> findById(BigInteger id);
+	Optional<User> findById(long id);
 
 	User insert(User user);
